pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "githubtakeout"
version = "0.1.17dev0"
description = "Backup and archive Git Repos and Gists from GitHub"
license = "MIT"
license-files = ["LICENSE"]
authors = [{name = "Corey Goldberg"}]
maintainers = [{name = "Corey Goldberg"}]
readme = "README.md"
requires-python = ">=3.12"
keywords = ["git", "github", "archive", "backup", "export", "takeout"]
classifiers = [
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "GitPython==3.1.45",
    "PyGithub==2.8.1",
    "python-dotenv==1.2.1",
    "rich==14.2.0",
]

[project.scripts]
githubtakeout = "githubtakeout:main"

[project.urls]
homepage = "https://github.com/cgoldberg/githubtakeout"
source = "https://github.com/cgoldberg/githubtakeout"
download = "https://pypi.org/project/githubtakeout"

[dependency-groups]
dev = [
    "tox",
]
lint = [
    "black",
    "refurb",
    "ruff",
]
test = [
    "pytest",
]
validate = [
    "packaging",
    "validate-pyproject",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.pytest.ini_options]
addopts = "-vv" # extra verbose
tmp_path_retention_policy = "none"

[tool.black]
line-length = 88
target-version = [
    "py312", "py313", "py314"
]

[tool.ruff]
line-length = 88
respect-gitignore = true
target-version = "py312"

[tool.ruff.lint]
select = [
    "D", "E", "F", "I", "PT", "RUF", "TID", "W"
]
extend-ignore = [
    "D100", "D101", "D102", "D103", "D104", "D105", "D106", "D107", "D203",
    "D205", "D211", "D213", "D400", "D401", "RUF012"
]
fixable = ["ALL"]

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 88

[tool.refurb]
python_version = "3.12"
enable_all = true
disable = [
    "FURB104", "FURB107", "FURB147", "FURB173", "FURB183", "FURB184"
]
